update test suite to be smarter about return codes

+++ gridftp/client/test/globus-ftp-client-caching-get-test.pl
@@ -35,24 +35,14 @@ sub basic_func
     my $tmpname = POSIX::tmpnam();
     my ($errors,$rc) = ("",0);
 
-    unlink('core');
-
     if($use_proxy == 0)
     {
         FtpTestLib::push_proxy("/dev/null");
     }
     
     my $command = "$test_exec -s gsiftp://$source_host$source_file >$tmpname 2>/dev/null";
-    $rc = run_command($command) / 256;
-    if(($use_proxy && $rc != 0) || (!$use_proxy && $rc == 0))
-    {
-        $errors .= "\n# Test exited with $rc. ";
-    }
-    if(-r 'core')
-    {
-        $errors .= "\n# Core file generated.";
-    }
-    if($use_proxy)
+    $errors = run_command($command, $use_proxy ? 0 : -1);
+    if($errors eq "" && $use_proxy)
     {
         my ($newtmp)=(POSIX::tmpnam());
 	system("cat '$local_copy' '$local_copy' > $newtmp");
@@ -85,24 +75,13 @@ push(@tests, "basic_func" . "(1);"); #Use proxy
 # Success if program returns 1 and no core file is generated.
 sub bad_url
 {
-    my $tmpname = POSIX::tmpnam();
     my ($errors,$rc) = ("",0);
     my ($bogus_url) = new Globus::URL("gsiftp://$source_host$source_file");
 
     $bogus_url->{path} = "/no-such-file-here";
-    unlink('core');
     
     my $command = "$test_exec -s ".$bogus_url->to_string()." >/dev/null 2>/dev/null";
-    $rc = run_command($command) / 256;
-    if($rc != 2)
-    {
-        $errors .= "\n# Test exited with $rc.";
-    }
-    if(-r 'core')
-    {
-        $errors .= "\n# Core file generated.";
-    }
-
+    $errors = run_command($command, 2);
     if($errors eq "")
     {
         ok('success', 'success');
@@ -112,7 +91,6 @@ sub bad_url
         $errors = "\n# Test failed\n# $command\n# " . $errors;
         ok($errors, 'success');
     }
-    unlink($tmpname);
 }
 push(@tests, "bad_url");
 
@@ -123,19 +101,11 @@ push(@tests, "bad_url");
 # a stronger measure of success here)
 sub abort_test
 {
-    my $tmpname = POSIX::tmpnam();
     my ($errors,$rc) = ("", 0);
     my ($abort_point) = shift;
 
-    unlink('core', $tmpname);
-
     my $command = "$test_exec -a $abort_point -s gsiftp://$source_host$source_file >/dev/null 2>/dev/null";
-    $rc = run_command($command) / 256;
-    if(-r 'core')
-    {
-        $errors .= "\n# Core file generated.";
-    }
-
+    $errors = run_command($command, -2);
     if($errors eq "")
     {
         ok('success', 'success');
@@ -145,7 +115,6 @@ sub abort_test
         $errors = "\n# Test failed\n# $command\n# " . $errors;
         ok($errors, 'success');
     }
-    unlink($tmpname);
 }
 for(my $i = 1; $i <= 41; $i++)
 {
@@ -163,24 +132,19 @@ sub restart_test
     my ($errors,$rc) = ("",0);
     my ($restart_point) = shift;
 
-    unlink('core', $tmpname);
+    unlink($tmpname);
 
     my $command = "$test_exec -r $restart_point -s gsiftp://$source_host$source_file >'$tmpname' 2>/dev/null";
-    $rc = run_command($command) / 256;
-    if($rc != 0)
+    $errors = run_command($command, 0);
+    if($errors eq "")
     {
-        $errors .= "\n# Test exited with $rc. ";
-    }
-    if(-r 'core')
-    {
-        $errors .= "\n# Core file generated.";
-    }
-    my ($newtmp)=(POSIX::tmpnam());
-    system("cat '$local_copy' '$local_copy' > $newtmp");
+        my ($newtmp)=(POSIX::tmpnam());
+        system("cat '$local_copy' '$local_copy' > $newtmp");
 
-    $errors .= compare_local_files($newtmp, $tmpname);
+        $errors .= compare_local_files($newtmp, $tmpname);
 
-    unlink($newtmp);	
+        unlink($newtmp);	
+    }
 
     if($errors eq "")
     {
